move tests from source to test area
